budget_project/requirements.txt
djangorestframework==3.14.0
Pillow==10.0.0
pytesseract==0.3.10
qrcode==7.4.2
firebase-admin==6.2.0
SpeechRecognition==3.10.0
pydub==0.25.1 